solution/requirements.txt
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
Django==5.1.5
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
